Add float32 array EclSum.to_numpy
This commit introduces pybind11 in order to accomplish the goal of
dumping float32 arrays. This could've been done by adding yet another C
function accessible from Python via cwrap. However, I made the decision
to have the Python API not depend on the C API directly, instead adding
Python bindings via pybind11. This allows us to extend the Python API
without exposing implementation details in C and avoids having the
Python API do manual memory management to deal with C.

+++ lib/ecl/ecl_unsmry_loader.cpp
@@ -44,13 +44,13 @@ unsmry_loader::~unsmry_loader() { ecl_file_close(file); }
 
 int unsmry_loader::length() const { return this->m_length; }
 
-std::vector<double> unsmry_loader::get_vector(int pos) const {
+std::vector<float> unsmry_loader::get_vector(int pos) const {
     if (pos >= size)
         throw std::out_of_range(
             "unsmry_loader::get_vector pos: " + std::to_string(pos) +
             " PARAMS_SIZE: " + std::to_string(size));
 
-    std::vector<double> data(this->length());
+    std::vector<float> data(this->length());
     int_vector_type *index_map = int_vector_alloc(1, pos);
     char buffer[4];
 
@@ -69,7 +69,7 @@ std::vector<double> unsmry_loader::get_vector(int pos) const {
 }
 
 // This is horribly inefficient
-double unsmry_loader::iget(int time_index, int params_index) const {
+float unsmry_loader::iget(int time_index, int params_index) const {
     int_vector_type *index_map = int_vector_alloc(1, params_index);
     float value;
     ecl_file_view_index_fload_kw(this->file_view, PARAMS_KW, time_index,
@@ -128,7 +128,7 @@ std::vector<int> unsmry_loader::report_steps(int offset) const {
 
 std::vector<time_t> unsmry_loader::sim_time() const {
     if (this->time_index >= 0) {
-        const std::vector<double> sim_seconds = this->sim_seconds();
+        const std::vector<float> sim_seconds = this->sim_seconds();
         std::vector<time_t> st(this->length(), this->sim_start);
 
         for (size_t i = 0; i < st.size(); i++)
@@ -150,16 +150,16 @@ std::vector<time_t> unsmry_loader::sim_time() const {
     }
 }
 
-std::vector<double> unsmry_loader::sim_seconds() const {
+std::vector<float> unsmry_loader::sim_seconds() const {
     if (this->time_index >= 0) {
-        std::vector<double> seconds = this->get_vector(this->time_index);
+        std::vector<float> seconds = this->get_vector(this->time_index);
         for (size_t i = 0; i < seconds.size(); i++)
             seconds[i] *= this->time_seconds;
 
         return seconds;
     } else {
         std::vector<time_t> st = this->sim_time();
-        std::vector<double> seconds(st.size());
+        std::vector<float> seconds(st.size());
 
         for (size_t i = 0; i < st.size(); i++)
             seconds[i] = util_difftime_seconds(this->sim_start, st[i]);
