+++ lib/private-include/detail/ecl/ecl_unsmry_loader.hpp
@@ -14,15 +14,15 @@ public:
                   int file_options);
     ~unsmry_loader();
 
-    std::vector<double> get_vector(int pos) const;
-    std::vector<double> sim_seconds() const;
+    std::vector<float> get_vector(int pos) const;
+    std::vector<float> sim_seconds() const;
     std::vector<time_t> sim_time() const;
     int length() const;
 
     time_t iget_sim_time(int time_index) const;
     double iget_sim_seconds(int time_index) const;
     std::vector<int> report_steps(int offset) const;
-    double iget(int time_index, int params_index) const;
+    float iget(int time_index, int params_index) const;
 
 private:
     int size; //Number of entries in the smspec index
